pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-graph-memory"
version = "0.1.0"
description = "Graph-enriched Obsidian memory for OpenClaw agents — GLiNER2 extraction + Gemini embeddings + ChromaDB + SQLite KG"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "google-generativeai>=0.8.0",
    "python-frontmatter>=1.1.0",
    "mcp[cli]>=1.0.0",
    "gliner2>=1.2.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
obsidian-memory = "obsidian_graph_memory.cli:main"
obsidian-memory-server = "obsidian_graph_memory.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_graph_memory"]

[tool.ruff]
line-length = 100
